rename datasources package to datasource
