+++ lib/data/repositories/product_repository_impl.dart
@@ -5,8 +5,8 @@ import '../../domain/repositories/product_repository.dart';
 import '../core/storage/storage_keys.dart';
 import '../core/storage/storage_service.dart';
 import '../core/utils/repository_call_handler.dart';
-import '../datasources/remote/favorites_remote_datasource.dart';
-import '../datasources/remote/product_remote_datasource.dart';
+import '../datasource/remote/favorites_remote_datasource.dart';
+import '../datasource/remote/product_remote_datasource.dart';
 import '../models/common/paginated_response.dart';
 import '../models/product_model.dart';
